Remove one layer and remove obsolete files
