--- nextjs/components/Pages/ChannelsPage/Channel.module.css
@@ -1,10 +1,0 @@
-.td {
-  padding: 0.75rem 1.5rem;
-  vertical-align: middle;
-}
-
-@media (min-width: 768px) {
-  .td {
-    max-width: 800px;
-  }
-}
